Add validation to CategoryService

+++ src/main/java/com/example/supermarket/service/CategoryService.java
@@ -1,21 +1,35 @@
 package com.example.supermarket.service;
 
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 
 import com.example.supermarket.entity.Category;
 import com.example.supermarket.repo.CategoryRepository;
+
+import jakarta.persistence.EntityNotFoundException;
 
 public class CategoryService {
 
     @Autowired
     private CategoryRepository categoryRepository;
 
+    public List<Category> findAll() {
+        List<Category> categories = categoryRepository.findAll();
+        if (categories.isEmpty()) {
+            throw new EntityNotFoundException("There are no categories");
+        }
+        return categories;
+    }
+
     public Category findById(int id) {
-        return categoryRepository.findById(id);
+        return categoryRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Category with id " + id + " not found"));
     }
 
     public Category findByName(String name) {
-        return categoryRepository.findByName(name);
+        return categoryRepository.findByName(name)
+                .orElseThrow(() -> new EntityNotFoundException("Categoy with name " + name + " not found"));
     }
 
     public void save(Category category) {
@@ -23,10 +37,16 @@ public class CategoryService {
     }
 
     public void deleteAll() {
+        if (findAll().isEmpty()) {
+            throw new EntityNotFoundException("There are no categories to delete");
+        }
         categoryRepository.deleteAll();
     }
 
     public void deleteById(int id) {
+        if (findById(id) == null) {
+            throw new EntityNotFoundException("There is no cartegory with id" + id + " to delete");
+        }
         categoryRepository.deleteById(id);
     }
 }
